+++ backend/src/main/java/io/swagger/service/StationService.java
@@ -2,6 +2,7 @@ package io.swagger.service;
 
 import io.swagger.model.Station;
 import io.swagger.repository.StationRepository;
+import io.swagger.repository.TrainRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -12,10 +13,12 @@ import java.util.Optional;
 public class StationService {
 
 	private final StationRepository stationRepository;
+	private final TrainRepository trainRepository;
 
 	@Autowired
-	public StationService(StationRepository stationRepository) {
+	public StationService(StationRepository stationRepository, TrainRepository trainRepository) {
 		this.stationRepository = stationRepository;
+		this.trainRepository = trainRepository;
 	}
 
 	public List<Station> getAllStations() {
@@ -36,7 +39,19 @@ public class StationService {
 		return stationRepository.save(station);
 	}
 
-	public void deleteStation(Integer stationId) {
-		stationRepository.deleteById(stationId);
+	public boolean deleteStation(Integer stationId) {
+		boolean isStationUsedInTrain = checkIfStationUsedInTrain(stationId);
+
+		if (isStationUsedInTrain) {
+			return false;
+		}
+		else {
+			stationRepository.deleteById(stationId);
+			return true;
+		}
+	}
+
+	private boolean checkIfStationUsedInTrain(Integer stationId) {
+		return trainRepository.existsBySourceStationStationIdOrDestinationStationStationId(stationId, stationId);
 	}
 }
